requirements.txt
click==8.1.7
jinja2==3.1.2
requests==2.31.0
reportlab==4.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
